+++ woodpecker-agent/cmd.sh
@@ -25,6 +25,43 @@ WOODPECKER_HOSTNAME="${WOODPECKER_HOSTNAME:-woodpecker-agent}"
 WOODPECKER_LOG_LEVEL="${WOODPECKER_LOG_LEVEL:-info}"
 WOODPECKER_SERVER="${WOODPECKER_SERVER:-}"
 
+import_ca_certificates() {
+    source_dir="/var/local/woodpecker-agent/certs"
+    destination_dir="/usr/local/share/ca-certificates"
+
+    if [ ! -z "$(ls ${source_dir}/)" ]; then
+        echo "* Importing CA certificate..."
+
+        for cert in $(ls -1 ${source_dir}/*); do
+            if $(openssl x509 -in "${cert}" -noout); then
+                echo "  Certificate: $(basename ${cert})"
+
+                datetime_valid=$(openssl x509 -in "${cert}" -noout --enddate | sed 's@notAfter=@@g')
+                epoch_valid=$(date --date="${datetime_valid}" --utc +'%s')
+                epoch_now=$(date --utc +'%s')
+
+                ca_certificate_fingerprint=$(openssl x509 -in "${cert}" -noout -fingerprint | sed 's@=@: @g')
+                ca_certificate_serial=$(openssl x509 -in "${cert}" -noout -serial | sed 's@serial=@Serial: @g')
+                ca_certificate_valid=$(openssl x509 -in "${cert}" -noout -enddate | sed 's@notAfter=@Valid Until: @g')
+
+                if [ "${epoch_valid}" -ge "${epoch_now}" ]; then
+                    echo "  ${ca_certificate_fingerprint}"
+                    echo "  ${ca_certificate_serial}"
+                    echo "  ${ca_certificate_valid}"
+                    cp -f "${cert}" "/usr/local/share/ca-certificates/"
+                    update-ca-certificates
+                    echo
+                else
+                    echo "  ${ca_certificate_valid}"
+                    echo "Error: CA certificate expired, exiting..."
+                    sleep 60
+                    exit 1
+                fi
+            fi
+        done
+    fi
+}
+
 assemble_command() {
     cmd=(exec)
     cmd+=(/usr/local/bin/woodpecker-agent)
@@ -129,6 +166,7 @@ assemble_command() {
 
 # Establish run order
 main() {
+    import_ca_certificates
     assemble_command
     "${cmd[@]}"
 }